requirements.txt
fastapi
uvicorn[standard]
pydantic
librosa
soundfile
transformers
datasets
torch
torchaudio
scipy
wandb
mlflow
numpy
dv
pandas
python-dotenv
tqdm
accelerate
evaluate
fairseq
audiomentations
jiwer
phonemizer
psycopg2-binary